refactor: MVC 패턴에 맞게 패키지 및 클래스 변경
