--- src/main/java/racing/Main.java
@@ -1,8 +1,0 @@
-package racing;
-
-public class Main {
-    public static void main(String[] args) {
-        RacingGame racingGame = new RacingGame();
-        racingGame.gameStart();
-    }
-}
