+++ src/main/java/racing/domain/RacingGame.java
@@ -1,31 +1,14 @@
-package racing;
+package racing.domain;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import static racing.utils.RandomNumber.makeOneRandomNumber;
-import static racing.view.Input.makeGameRepeatCount;
-import static racing.view.Input.makeCarNames;
-import static racing.view.Output.*;
 
 public class RacingGame {
     public static String SEPARATOR = ",";
 
-    public void gameStart() {
-        printStartMessage();
-        String carNames = makeCarNames();
-        printInputCountMessage();
-        int count = makeGameRepeatCount();
-        printResultMessage();
-
-        String[] carNameArray = carNames.split(SEPARATOR);
-        List<Car> carList = makeCars(carNameArray);
-
-        repeatMoveCars(count, carList);
-        printWinMessage(findWinner(carList));
-    }
-
-    private List<Car> makeCars(String[] carNameArray) {
+    public List<Car> makeCars(String[] carNameArray) {
         List<Car> cars = new ArrayList<>();
 
         for (String carName : carNameArray) {
@@ -34,20 +17,14 @@ public class RacingGame {
         return cars;
     }
 
-    private void repeatMoveCars(int count, List<Car> carList) {
+    public List<Car> repeatMoveCars(int count, List<Car> carList) {
         for (int i = 0; i < count; i++) {
-            moveCars(carList);
+            carList = moveCars(carList);
         }
+        return carList;
     }
 
-    private void moveCars(List<Car> carList) {
-        for (int i = 0; i < carList.size(); i++) {
-            carList.get(i).move(makeOneRandomNumber());
-        }
-        printNowDistance(carList);
-    }
-
-    private List<String> findWinner(List<Car> carList) {
+    public List<String> findWinner(List<Car> carList) {
         List<String> winnerList = new ArrayList<>();
         int winnerCondition = findWinnerCondition(carList);
         for (Car car : carList) {
@@ -56,6 +33,13 @@ public class RacingGame {
             }
         }
         return winnerList;
+    }
+
+    private List<Car> moveCars(List<Car> carList) {
+        for (int i = 0; i < carList.size(); i++) {
+            carList.get(i).move(makeOneRandomNumber());
+        }
+        return carList;
     }
 
     private int findWinnerCondition(List<Car> carList) {
