+++ src/main/java/racing/view/Output.java
@@ -1,6 +1,6 @@
 package racing.view;
 
-import racing.Car;
+import racing.domain.Car;
 
 import java.util.List;
